Amended the tests, but they are failing because the test score is not always being stored with the answers
